+++ libs/InteractiveToolkit-AppKit/opengl/appkit-gl-engine/src/Components/ComponentCameraMove.cpp
@@ -202,7 +202,7 @@ namespace AppKit
                 writer.EndObject();
                 
             }
-            void ComponentCameraMove::Deserialize(rapidjson::Value &_value, std::unordered_map<intptr_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<intptr_t, std::shared_ptr<Component>> &component_map){
+            void ComponentCameraMove::Deserialize(rapidjson::Value &_value, std::unordered_map<uint64_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<uint64_t, std::shared_ptr<Component>> &component_map){
                 if (!_value.HasMember("type") || !_value["type"].IsString())
                     return;
                 if (!strcmp(_value["type"].GetString(), ComponentCameraMove::Type) == 0)
