+++ src/tech-demo-animation/components/RotatingCross.cpp
@@ -53,7 +53,7 @@ namespace AppKit
                 writer.EndObject();
                 
             }
-            void RotatingCross::Deserialize(rapidjson::Value &_value, std::unordered_map<intptr_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<intptr_t, std::shared_ptr<Component>> &component_map){
+            void RotatingCross::Deserialize(rapidjson::Value &_value, std::unordered_map<uint64_t, std::shared_ptr<Transform>> &transform_map, std::unordered_map<uint64_t, std::shared_ptr<Component>> &component_map){
                 if (!_value.HasMember("type") || !_value["type"].IsString())
                     return;
                 if (!strcmp(_value["type"].GetString(), RotatingCross::Type) == 0)
